fix (comments): *fix: a bug when publishing with gen key due to different created_at values *clean-up code
